add liquidity chart
replace candles feed with quoteFeed factory that supports
both price (candles) and liquidity
